requirements.txt
alembic==1.1.0
aniso8601==7.0.0
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.4
decorator==4.4.0
Flask==1.1.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.21
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
more-itertools==7.2.0
numpy==1.17.2
packaging==19.1
paho-mqtt==1.4.0
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.8
sqlalchemy-seeder==0.3.0
style==1.1.6
traitlets==4.3.2
update==0.0.1
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
